LayoutTests/imported/w3c/web-platform-tests/fetch/corb/style-css-mislabeled-as-html-nosniff.sub-expected.txt

CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/fetch/corb/resources/css-mislabeled-as-html-nosniff.css
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/fetch/corb/resources/css-mislabeled-as-html-nosniff.css
Header example

Paragraph body


PASS Untitled
